==> STEP2_EDA.R <==
setwd("E:/school/data mining/project/mimic-iii-clinical-database-1.4/mimic-iii-clinical-database-1.4/codes")
source("Global.R")
pt <-read.csv("PATIENTS.csv")
biodata <-read.csv("cardiacSyndromes.csv")
service <- read.csv("serviceGroup.csv") %>% 
  select(where(is.numeric) ,-X)
lab <- read.csv("labGroup.csv") %>% 
  select(where(is.numeric),-X)
med <- read.csv("mediGroup.csv") %>% 
  select(where(is.numeric),-X)
microb <- read.csv("microGroup.csv") %>% 
  select(where(is.numeric),-X)
proc <- read.csv("procGroup.csv") %>% 
  select(where(is.numeric),-X)

data <-biodata %>% 
  mutate_all(na_if,"") %>% 
  mutate(MARITAL_STATUS = if_else(is.na(MARITAL_STATUS), "UNKNOWN", MARITAL_STATUS)) 
data$MARITAL_STATUS<-trimws(data$MARITAL_STATUS)
data$MARITAL_STATUS<-ifelse(grepl("(DEFAULT)",data$MARITAL_STATUS),"UNKNOWN",data$MARITAL_STATUS)

#data$MARITAL_STATUS<-ifelse(grepl("UNKNOWN (DEFAULT)",data$MARITAL_STATUS),"UNKNOWN",data$MARITAL_STATUS)

data$ETHNICITY<-ifelse(grepl("UNKNOWN",data$ETHNICITY),"UNKNOWN",data$ETHNICITY)
data$ETHNICITY<-ifelse(grepl("PATIENT DECLINED",data$ETHNICITY),"UNKNOWN",data$ETHNICITY)
data$Outcome <- ifelse(data$EXPIRE_FLAG==1,"Deceased", "Survived")
data$ETHNICITY2<-ifelse(grepl("ASIAN",data$ETHNICITY),"ASIAN",data$ETHNICITY)
data$ETHNICITY2<-ifelse(grepl("MIDDLE EASTERN",data$ETHNICITY2),"ASIAN",data$ETHNICITY2)
data$ETHNICITY2<-ifelse(grepl("HISPANIC",data$ETHNICITY2),"HISPANIC",data$ETHNICITY2)
data$ETHNICITY2<-ifelse(grepl("SOUTH AMERICAN",data$ETHNICITY2),"HISPANIC",data$ETHNICITY2)
data$ETHNICITY2<-ifelse(grepl("WHITE - BRAZILIAN",data$ETHNICITY2),"HISPANIC",data$ETHNICITY2)
data$ETHNICITY2<-ifelse(grepl("CARIBBEAN ISLAND",data$ETHNICITY2),"HISPANIC",data$ETHNICITY2)
data$ETHNICITY2<-ifelse(grepl("ALASKA NATIVE",data$ETHNICITY2),"CAUCASIAN",data$ETHNICITY2)
data$ETHNICITY2<-ifelse(grepl("NATIVE HAWAIIAN",data$ETHNICITY2),"CAUCASIAN",data$ETHNICITY2)
data$ETHNICITY2<-ifelse(grepl("PORTUGUESE",data$ETHNICITY2),"CAUCASIAN",data$ETHNICITY2)
data$ETHNICITY2<-ifelse(grepl("WHITE",data$ETHNICITY2),"CAUCASIAN",data$ETHNICITY2)
data$ETHNICITY2<-ifelse(grepl("BLACK",data$ETHNICITY2),"BLACK",data$ETHNICITY2)
data$ETHNICITY2<-ifelse(grepl("MULTI RACE ETHNICITY",data$ETHNICITY2),"OTHER",data$ETHNICITY2)
data$ETHNICITY2<-ifelse(grepl("UNABLE TO  cannot open the connection
In addition: Warning message:
 OBTAIN",data$ETHNICITY2),"UNKNOWN",data$ETHNICITY2)
data$Outcome <-data$EXPIRE_FLAG


Table1Data<-data %>% 
  select(ADMISSION_TYPE,ADMISSION_LOCATION,INSURANCE,LANGUAGE,RELIGION,MARITAL_STATUS,ETHNICITY2,Outcome,GENDER,AGE,nAdmissions)


#write.csv(Table1Data, "Table1Data.csv")
## Time PLot

dataPlot <-biodata %>% 
  select(ADMITTIME,EXPIRE_FLAG) %>% 
  mutate(year = lubridate::year(ADMITTIME)) %>% 
  select(-ADMITTIME)

Dead<-dataPlot %>% 
  filter(EXPIRE_FLAG==0) %>% 
  count(year, sort = TRUE) %>% 
  mutate(status = "Survived")

survive<-dataPlot %>% 
  filter(EXPIRE_FLAG== 1) %>% 
  count(year, sort = TRUE) %>% 
  mutate(status = "Dead")

dataPlot1<-rbind(Dead,survive)

dataPlot1

dataPlot<-ggplot(dataPlot1, aes(x = year, y = n)) + 
  geom_line(aes(color = status), size = 1) +
  #scale_color_manual(values = c("#00AFBB", "#E7B800")) +
  theme_economist() +
  ggtitle("From 2100 to 2190 (Full Data Set)") +
  xlab("Year(shifted)") + ylab("Number of admissions")



p2 <- dataPlot1 %>%
  #filter(year<2130) %>%
  ggplot(aes(year, n)) +
  #geom_line(color = palette_light()[[1]], alpha = 0.5) +
  geom_line(aes(color = status), size = 1) +
  #geom_point(color = palette_light()[[1]]) +
  #geom_smooth(method = "loess", span = 0.2, se = FALSE) +
  #theme_economist() +
  theme_wsj()+
  labs(
    title = "2100 to 2130 (Zoomed In To Show Cycle)",
    caption = "Number of admissions"
  )
